backend/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
python-dotenv==1.0.0
transformers==4.30.2
torch==2.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.3
safetensors==0.3.1
flask-bcrypt==1.0.1